Data Projects/Fuel Price Analytics/Dashboard/requirements.txt
dash==2.17.0
dash-bootstrap-components==1.6.0
pandas==2.2.2
pillow==10.3.0